requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
slowapi==0.1.9
python-dotenv==1.0.1
